goback: update `clean` command to remove by backup type
If there are, say, 10 backups of each type (10 daily, 10 weekly, 10
monthly), and the user wants to keep the latest 5 backups, it should
delete 15 backups to keep 5 of each, not necessarily just the last 5 or
something to that effect. That way, older monthly backups can be kept
around much longer than the daily backup logs.